projects/progo-protein-evaluation-tm-plddt/requirements.txt
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
tqdm>=4.65.0
biopython>=1.81
scikit-learn>=1.3.0
pyyaml>=6.0.1
jsonschema>=4.17.3
biotite>=0.38.0 